Update Jenkinsfile
In this updated version, a cache directory .cache/pip is created within your workspace using mkdir -p .cache/pip. This directory will be writable by the user running the pipeline. Then, when installing the requirements with pip, the --cache-dir option is used to specify the cache directory as .cache/pip.