Update Jenkinsfile
In this updated version, we create the virtual environment using the --clear option to ensure a clean environment. We then set the pip cache directory to /tmp/pip-cache, which is typically writable. Finally, we install the dependencies using pip install -r requirements.txt. This should resolve the permission issue with the cache directory.